backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
openai==1.51.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-CORS==5.0.0
